requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
transformers==4.37.2
torch==2.2.0
python-pptx==0.6.23
detectron2==0.6